refactor larger pieces into smaller components
