+++ app/lib/metadata.ts
@@ -1,0 +1,9 @@
+import { Metadata } from "next";
+
+export const BASE_META_DATA: Metadata = {
+  title: {
+    template: "%s | Danny Grimmig",
+    default: "Danny Grimmig",
+  },
+  description: "welcome to my portfolio site.",
+};
