+++ app/ui/NavBar.tsx
@@ -1,15 +1,28 @@
+"use client";
+
 import Link from "next/link";
+import { usePathname } from "next/navigation";
 
 export function NavBar() {
+  const pathname = usePathname();
+
   return (
     <nav className="z-10 font-mono sticky top-0 bg-white border-b border-black py-4 px-2 sm:px-8 flex gap-16 items-end">
       <Link href="/">
         <h1 className="text-4xl">Danny&#128075;</h1>
       </Link>
+
       <ul className="flex gap-4 sm:gap-8">
         {LINKS.map((link) => (
           <Link key={link} href={`/${link}`}>
-            {link}
+            <p
+              className={`${
+                pathname === `/${link}` &&
+                "underline underline-offset-4	decoration-1"
+              }`}
+            >
+              {link}
+            </p>
           </Link>
         ))}
       </ul>
